Optimization of the HGCDoublet data structure

+++ RecoHGCal/TICL/plugins/HGCDoublet.h
@@ -32,21 +32,21 @@ public:
         outerZ_((*layerClusters)[outerClusterId].z()),
         alreadyVisited_(false) {}
 
-  double innerX() const { return innerX_; }
+  float innerX() const { return innerX_; }
 
-  double outerX() const { return outerX_; }
+  float outerX() const { return outerX_; }
 
-  double innerY() const { return innerY_; }
+  float innerY() const { return innerY_; }
 
-  double outerY() const { return outerY_; }
+  float outerY() const { return outerY_; }
 
-  double innerZ() const { return innerZ_; }
+  float innerZ() const { return innerZ_; }
 
-  double outerZ() const { return outerZ_; }
+  float outerZ() const { return outerZ_; }
 
-  double innerR() const { return innerR_; }
+  float innerR() const { return innerR_; }
 
-  double outerR() const { return outerZ_; }
+  float outerR() const { return outerZ_; }
 
   int innerClusterId() const { return innerClusterId_; }
 
@@ -59,12 +59,12 @@ public:
   bool checkCompatibilityAndTag(
       std::vector<HGCDoublet> &, const std::vector<int> &, float, float minCosPointing = 1., bool debug = false);
 
-  int areAligned(double xi,
-                 double yi,
-                 double zi,
-                 double xo,
-                 double yo,
-                 double zo,
+  int areAligned(float xi,
+                 float yi,
+                 float zi,
+                 float xo,
+                 float yo,
+                 float zo,
                  float minCosTheta,
                  float minCosPointing,
                  bool debug = false) const;
@@ -80,14 +80,14 @@ private:
   const int innerClusterId_;
   const int outerClusterId_;
 
-  const double innerR_;
-  const double outerR_;
-  const double innerX_;
-  const double outerX_;
-  const double innerY_;
-  const double outerY_;
-  const double innerZ_;
-  const double outerZ_;
+  const float innerR_;
+  const float outerR_;
+  const float innerX_;
+  const float outerX_;
+  const float innerY_;
+  const float outerY_;
+  const float innerZ_;
+  const float outerZ_;
   bool alreadyVisited_;
 };
